+++ miamoremustachio/app/main.js
@@ -13,43 +13,18 @@ const {
   idGenerator,
 } = require('./modules/helpers');
 
-const getId = () => idGenerator.next().value;
+const { validationLayer } = require('./modules/validation');
 
 const { TITLE, STATUS, PRIORITY } = TASK_PROPERTIES;
 const { TO_DO, IN_PROGRESS, DONE } = STATUSES;
 const { LOW, MEDIUM, HIGH } = PRIORITIES;
 const { TASK_NOT_FOUND, INVALID_TITLE, INVALID_STATUS, INVALID_PRIORITY } = ERRORS;
 
-function isTitleValid(title) {
-  return (typeof title === 'string' && title.length < MAX_TITLE_LENGTH);
-}
+const { fn: isTitleValid } = validationLayer[TITLE];
+const { fn: isStatusValid } = validationLayer[STATUS];
+const { fn: isPriorityValid } = validationLayer[PRIORITY];
 
-function isStatusValid(status) {
-  const validStatusesList = Object.values(STATUSES);
-  
-  return validStatusesList.includes(status);
-}
-
-function isPriorityValid(priority) {
-  const validPrioritiesList = Object.values(PRIORITIES);
-
-  return validPrioritiesList.includes(priority);
-}
-
-const validationLayer = {
-  [TITLE]: {
-    fn: isTitleValid,
-    errorMessage: INVALID_TITLE, 
-  },
-  [STATUS]: {
-    fn: isStatusValid,
-    errorMessage: INVALID_STATUS,
-  },
-  [PRIORITY]: {
-    fn: isPriorityValid,
-    errorMessage: INVALID_PRIORITY, 
-  },
-};
+const getId = () => idGenerator.next().value;
 
 const toDo = {
   list: [
